feat(web): merge the plan panel into a three-tab bluesky panel
Composing a scan, watching the queue and reading a run's data were split
across two rail entries, and the BLUESKY panel had to open by explaining
where plans were composed. They are now one panel with Plans, Queue and
Results tabs, contributed to the tile bar as a view switcher.

Queue state and both halts — stop after current item, abort running plan —
stay on screen across all three tabs rather than living behind one. They must
be live from first paint in every mode, and the bar that carries the tab
strip collapses in Simple. Start stays inside the Queue tab, where arming a
queue is a deliberate act. Selecting a run opens it under Results, replacing
the cross-panel run_id handoff the split required.

The sidecar serves the merged bundle at /plan/ as well as /bluesky/, so a
deployment whose config still registers a plan panel keeps working for one
release; the panel id also had to be retargeted in the build injectors, the
draft tool and the activity strip's suppression table.

The scan panel's visual baselines are dropped rather than updated: they have
to be regenerated on Linux, which this change cannot do.
