+++ CHANGELOG.md
@@ -31,6 +31,19 @@ Compatibility is documented in release notes, not encoded in the version string.
   lattice summary stats moved into the panel body, Channel Finder's pipeline
   switcher and corpus stats into a bottom strip, and the lattice Baseline
   button now asks for confirmation before overwriting.
+- The rest of the docked panels moved their toolbars into that one header bar
+  too. **WORKSPACE** contributes its filter, its Types/Activity switch and its
+  ⋯ menu; **KNOWLEDGE** its search; **EVENTS** its Activity/Triggers tabs. A
+  panel's search box now renders with the same magnifier as the terminal's own
+  search, so the two read alike. In Simple view the search stays in the panel
+  body, where that view puts it front and centre.
+- The **PLAN** and **BLUESKY** tabs are now one **BLUESKY** panel with three
+  views — Plans, Queue, Results. The queue's state and its two halts (**Stop
+  after current item**, **Abort running plan**) stay on screen across all
+  three, and picking a run in Queue opens it under Results. Projects that
+  still register a `plan` panel keep working for one release: the sidecar
+  serves the merged panel at `/plan/` too. Drop `plan` from your profile's
+  `web_panels` and remove any `web.panels.plan.*` override.
 - `osprey -v` (`--verbose`) shows debug output, including every container
   command a deploy runs. Normal runs no longer echo those commands, so a
   deploy reads as a report — ending in the endpoint summary — rather than a
@@ -305,6 +318,8 @@ Compatibility is documented in release notes, not encoded in the version string.
   change. Also removed dead panel chrome the audit surfaced: ARIEL's unwired
   "Connected" indicator and the System Health panel's no-op manual refresh
   and misleading fetch-time timestamp.
+- The **EVENTS** panel drew two header bars when docked — its own, plus the
+  tile's. It now hides its own, like every other panel.
 - On Docker Desktop (macOS/Windows), `osprey deploy up` now repairs a web
   stack that is fully healthy yet unreachable from the browser. Docker
   Desktop forwards a host-network port only if it watched the container open
