+++ docs/source/how-to/web-terminal/panels.rst
@@ -50,36 +50,40 @@ the entries for you):
 The hub shows the service as a tab and proxies requests to it from the same
 origin, so the browser never needs direct access to the backing port.
 
-The Bluesky panels
-------------------
+The Bluesky panel
+-----------------
 
-Two tabs cover a scan end to end, both served by the ``bluesky-panels``
-sidecar:
+One tab covers a scan end to end, served by the ``bluesky-panels`` sidecar.
+**BLUESKY** (``/bluesky/``) has three views, and the queue's state — with
+**Stop after current item** and **Abort running plan** — stays on screen across
+all three:
 
-**PLAN** (``/plan/``) is where a plan is composed. It binds to the same shared
-draft the OSPREY agent edits, so a field the agent sets glows in the form as it
-lands, and a field you change by hand flows back to the agent. **Add to queue**
-puts the exact revision on screen into the scan queue.
+**Plans** is where a plan is composed. It binds to the same shared draft the
+OSPREY agent edits, so a field the agent sets glows in the form as it lands,
+and a field you change by hand flows back to the agent. **Add to queue** puts
+the exact revision on screen into the scan queue.
 
-**BLUESKY** (``/bluesky/``) is the queue and the results. Its top half lists
-what the queue server is holding, with **Start queue**, **Stop after current
-item** and **Abort running plan**; its bottom half shows the selected run's
-table and live chart. :doc:`/how-to/bluesky/queue` covers what those controls
+**Queue** lists what the queue server is holding, with **Start queue** and the
+reorder/remove controls, plus the runs that have finished. Picking any run
+opens it under Results. :doc:`/how-to/bluesky/queue` covers what those controls
 do.
+
+**Results** shows the selected run's record, table and live chart.
 
 .. note::
 
-   **BLUESKY was called RESULTS**, and it now holds the queue as well as the
-   results the old tab showed. The sidecar still serves the same bundle at
-   ``/results/`` **for one release**, so existing bookmarks and panel entries
-   keep resolving; the alias goes away after that, so move any
-   ``web.panels.results.*`` entry of your own to ``web.panels.bluesky.*``.
+   **PLAN and RESULTS were separate tabs.** RESULTS became BLUESKY when that
+   bundle absorbed the queue, and PLAN then became BLUESKY's Plans view. The
+   sidecar still serves the same bundle at ``/results/`` and ``/plan/`` **for
+   one release**, so existing bookmarks and panel entries keep resolving; the
+   aliases go away after that, so drop any ``web.panels.plan.*`` entry of your
+   own and move ``web.panels.results.*`` to ``web.panels.bluesky.*``.
 
-   In the ``control-assistant`` preset that rename is already made. It changes
-   the preset's resolved content, so an **already-deployed project will report
-   staleness** on its next ``osprey deploy up``. That advisory is correct
-   rather than noise — unlike a cosmetic re-pin, this one renames a tab a user
-   sees — and the remedy is the rebuild the advisory prints.
+   In the ``control-assistant`` preset both changes are already made. They
+   change the preset's resolved content, so an **already-deployed project will
+   report staleness** on its next ``osprey deploy up``. That advisory is
+   correct rather than noise — unlike a cosmetic re-pin, this one changes the
+   tabs a user sees — and the remedy is the rebuild the advisory prints.
 
 Adding your own panel
 ---------------------
