--- src/osprey/interfaces/bluesky_panels/panels/plan/manifest.json
@@ -1,1 +1,0 @@
-{"id":"plan","label":"PLAN","entry":"index.html","version":1}
