+++ src/osprey/mcp_server/bluesky/tools/draft.py
@@ -2,7 +2,7 @@
 
 The draft is the ONE shared, live-editable staging surface — a draft of the
 next run. The bridge holds a single server-side draft (``{plan_name,
-plan_args, revision, ...}``) that the agent and the human's plan panel both
+plan_args, revision, ...}``) that the agent and the human's BLUESKY panel both
 edit. These three tools are thin HTTP clients of that draft — they never touch
 hardware, never require arming, and never pass through an approval prompt:
 editing the draft only stages what a future ``queue_add`` (or the human's
@@ -22,7 +22,7 @@ Same conventions as the other tool modules: ``async def``, JSON string return
 from ``bluesky/server_context.py`` for translating a non-2xx bridge response.
 
 Every write this module makes carries a fixed ``client_id: "mcp-agent"`` so
-the bridge's SSE frames (and the human's plan panel) can distinguish agent
+the bridge's SSE frames (and the human's BLUESKY panel) can distinguish agent
 edits from the human's own, and so the panel's echo-suppression never
 swallows an agent edit.
 """
@@ -48,19 +48,22 @@ logger = logging.getLogger("osprey.mcp_server.bluesky.tools.draft")
 
 _CLIENT_ID = "mcp-agent"
 
-# Canonical id the build registers the plan panel under (web.panels.plan) and
-# the sidecar path segment it is always mounted at (panels/plan -> /plan).
-_DEFAULT_PLANS_PANEL_ID = "plan"
-_PLAN_PANEL_PATH_SEGMENT = "plan"
+# Canonical id the build registers the panel under (web.panels.bluesky) and the
+# sidecar path segments it is served at, in preference order. ``plan`` is the
+# pre-merge mount of what is now the panel's Plans tab; the sidecar still serves
+# the same bundle there, so a deployment whose config.yml predates the merge is
+# still found — but only after the current spelling has been looked for.
+_DEFAULT_PLANS_PANEL_ID = "bluesky"
+_PLANS_PANEL_PATH_SEGMENTS = ("bluesky", "plan")
 
 
 def _plans_panel_id() -> str:
-    """Resolve the web-terminal panel id of the human's plan panel.
+    """Resolve the web-terminal panel id of the human's BLUESKY panel.
 
-    A successful draft edit is highlighted on the plan panel's rail entry, so
-    the emit must carry the id the web terminal actually knows the panel by:
-    the ``web.panels.<id>`` mapping key. The build registers the panel with
-    the canonical id ``plan`` mounted at the sidecar path ``/plan/``; a
+    A successful draft edit is highlighted on that panel's rail entry, so the
+    emit must carry the id the web terminal actually knows the panel by: the
+    ``web.panels.<id>`` mapping key. The build registers the panel with the
+    canonical id ``bluesky`` mounted at the sidecar path ``/bluesky/``; a
     facility that registered it under a different id is found by that fixed
     mount path. Same config fallback chain as the bridge URL resolution
     (``osprey.bluesky_bridge_connection.resolve_bridge_url``): config.yml via
@@ -70,11 +73,13 @@ def _plans_panel_id() -> str:
     from osprey.utils.workspace import load_osprey_config
 
     panels = load_osprey_config().get("web", {}).get("panels", {})
-    if isinstance(panels, dict):
+    if not isinstance(panels, dict):
+        return _DEFAULT_PLANS_PANEL_ID
+    for segment in _PLANS_PANEL_PATH_SEGMENTS:
         for panel_id, spec in panels.items():
             if not isinstance(spec, dict):
                 continue
-            if str(spec.get("path", "")).strip("/").split("/")[0] == _PLAN_PANEL_PATH_SEGMENT:
+            if str(spec.get("path", "")).strip("/").split("/")[0] == segment:
                 return str(panel_id)
     return _DEFAULT_PLANS_PANEL_ID
 
@@ -83,7 +88,7 @@ def _notify_draft_activity(tool: str, detail: str | None) -> None:
     """Sync body of the fire-and-forget activity emit (worker thread only).
 
     Reports a successful draft edit to the Web Terminal so the UI can
-    highlight the plan panel (``notify_agent_activity`` is blocking, hence
+    highlight the BLUESKY panel (``notify_agent_activity`` is blocking, hence
     dispatched via ``anyio.to_thread.run_sync``; it swallows all exceptions
     itself — a missing web terminal never affects the tool result). Resolving
     the panel id here keeps the config read off the event loop alongside the
@@ -99,8 +104,9 @@ def _notify_draft_activity(tool: str, detail: str | None) -> None:
 async def get_draft() -> str:
     """Read the shared plan draft. Reaches NO hardware.
 
-    The draft is the server-held scratch state the agent and the human's plan
-    panel both edit — this only reads it back, it never mutates anything.
+    The draft is the server-held scratch state the agent and the human's
+    BLUESKY panel both edit — this only reads it back, it never mutates
+    anything.
 
     Returns:
         JSON ``{"draft", "revision"}``. ``draft`` is ``null`` when no draft
@@ -127,13 +133,14 @@ async def set_draft(
     """Create or edit the shared plan draft — the staging surface for the next run.
 
     The draft is the ONE shared, live-editable surface both you and the human
-    fill before a run is launched. This edit fills the human's PLAN panel live:
-    every open panel reflects it within about a second and flashes exactly the
-    fields whose values changed — the bridge computes ``changed[]`` by comparing
-    values, so re-sending an already-current value is a silent no-op (no flash,
-    no revision bump). Setting ``plan_name`` on a draft that already names a
-    different plan replaces ``plan_args`` (with ``plan_args_patch``'s contents,
-    if also given); setting ``plan_name`` when no draft exists creates one.
+    fill before a run is launched. This edit fills the human's BLUESKY panel
+    live (its Plans view): every open panel reflects it within about a second
+    and flashes exactly the fields whose values changed — the bridge computes
+    ``changed[]`` by comparing values, so re-sending an already-current value is
+    a silent no-op (no flash, no revision bump). Setting ``plan_name`` on a
+    draft that already names a different plan replaces ``plan_args`` (with
+    ``plan_args_patch``'s contents, if also given); setting ``plan_name`` when
+    no draft exists creates one.
     Prefer one complete set_draft call (``plan_name`` plus the full
     ``plan_args_patch``) over a trickle of partial edits, so the human sees a
     coherent draft rather than a half-filled form.
@@ -213,7 +220,7 @@ async def clear_draft() -> str:
     """Clear the shared plan draft. Reaches NO hardware. Idempotent. Destructive.
 
     Wipes the ONE shared staging surface — the same draft the human may be
-    reviewing or filling in their plan panel right now — back to empty, and
+    reviewing or filling in their BLUESKY panel right now — back to empty, and
     bumps the revision. Use it deliberately; do not clear a draft just to
     start over when set_draft can replace ``plan_name`` in place. The sole
     clear path (there is no ``clear`` flag on set_draft). Either the agent or
